+++ NEWS.md
@@ -1,4 +1,22 @@
-﻿# legislatoR 0.2.0
+# legislatoR 1.0
+  
+* All legislative periods of Spain's Congreso de los Diputados added to the database (14 in total).
+* Ongoing session of the Austrian Nationalrat added to the database (27th) and concluded session updated (26th).
+* Ongoing session of the Canadian House of Commons added to the database (43th) and concluded session updated (42th).
+* Ongoing session of the Irish Dáil added to the database (33th) and concluded session updated (32th).
+* Ongoing session of the UK House of Commons added to the database (58th) and concluded session updated (57th).
+* Irish Dáil data integrated with Database of Parliamentary Speeches in Ireland.
+* Czech Poslanecka Snemovna data integrated with ParlSpeech V1.
+* UK House of Commons data integrated with ParlSpeech V1.
+* Spanish Congreso de los Diputados data integrated with ParlSpeech V2.
+* Austrian Nationalrat data integrated with ParlSpeech V2.
+* Daily Wikipedia page traffic now goes back to December 2007 (except for the US House, where file storage restrictions currently limit traffic to range back to July 2009 only).
+* New `cld_content` function that offers a quick overview of the CLD's scope and valid three-letter country codes, and helps to conventiently loop/map over legislatures and sessions.
+* New Logo/Hexsticker.
+* "Introducing legislatoR" Vignette added.
+
+
+# legislatoR 0.2.0
 
 * 1st to 37th legislative periods of UK's House of Commons added to the database (37 in total).
 * Missing legislators and services variable added to all sessions of the French Assemblée.